telegram_bridge/requirements.txt
blinker==1.9.0
click==8.3.2
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyaes==1.6.1
pyasn1==0.6.3
python-dotenv==1.2.2
rsa==4.9.1
Telethon==1.42.0
Werkzeug==3.1.8
pymysql
openai
